scripts/mining/processor/requirements.txt
boto3==1.35.39
python-dotenv==1.0.1
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
oauth2client==4.1.3
pandas==2.2.2